pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "curator"
version = "1.0.0rc1"
description = "Content-aware artifact intelligence layer for files"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Jake Leese"},
]
keywords = ["files", "deduplication", "lineage", "metadata", "curation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Filesystems",
    "Topic :: Utilities",
]

dependencies = [
    # Core data layer
    "pydantic>=2.5",
    "platformdirs>=4.0",
    "tomli-w>=1.0",  # tomli is stdlib (tomllib) for 3.11+; only need writer

    # Hashing
    "xxhash>=3.4",
    # ppdeep is vendored under curator/_vendored/ — no pip dep
    # send2trash is vendored under curator/_vendored/ — no pip dep

    # File-type detection
    "filetype>=1.2",

    # Plugin framework
    "pluggy>=1.3",

    # Logging & UI
    "loguru>=0.7",
    "rich>=13.7",
    "typer>=0.12",
    "questionary>=2.0",

    # File watching (Phase Beta but harmless to require now)
    "watchfiles>=0.21",

    # REST + HTTP (Phase Gamma but available for plugin authors)
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-qt>=4.2",
    "hypothesis>=6.95",
    "ruff>=0.3",
    "mypy>=1.8",
]

beta = [
    # Phase Beta optional plugins (install when needed)
    "apsw>=3.45",
    "datasketch>=1.6",         # MinHash + MinHashLSH for v0.13+ fuzzy candidate index
    "imagehash>=4.3",
    "python-magic>=0.4",
    "pypdf>=4.0",
    "tantivy>=0.21",
    "tree-sitter-python>=0.21",
    "watchfiles>=0.21",        # Tier 6 reactive watcher for v0.16+
]

cloud = [
    # Cloud source plugins (Phase Beta+)
    "PyDrive2>=1.20",            # Google Drive
    "msgraph-sdk>=1.0",          # OneDrive (Microsoft Graph)
    "dropbox>=12.0.2",           # Dropbox (pin >=12.0.2 for Jan 2026 cert compat)
]

organize = [
    # Phase Gamma: smart drive organizer + cleanup utilities (F1, F6–F17)
    "psutil>=5.9",               # Open-handle detection for SafetyService
    "mutagen>=1.47",             # Audio tag reading (F2 music)
    "tinytag>=2.0",              # Lighter alternative for fast bulk audio scans
    "Pillow>=10.0",              # EXIF reading (F3 photos)
    "pypdf>=4.0",                # PDF metadata reading (F4 documents)
    "musicbrainzngs>=0.7",       # MusicBrainz lookups for music with empty tags (F2 enrichment)
]

windows = [
    # Windows-specific (Phase Gamma)
    "pywin32>=306",
    "pystray>=0.19",
    "APScheduler>=3.10",
]

gui = [
    # Phase Beta gate 4: native PySide6 desktop GUI (v0.34)
    "PySide6>=6.6",
    # v0.41 Lineage Graph view: graph algorithms (layout, traversal)
    "networkx>=3.0",
]

all = [
    "curator[dev,beta,cloud,organize,windows,gui]",
]

[project.scripts]
curator = "curator.cli.main:app"

[project.entry-points.curator]
# Built-in plugins are loaded via curator.plugins.core.register_core_plugins()
# External plugins register here as they're installed:
# example_plugin = "curatorplug.example:Plugin"

[project.urls]
Homepage = "https://github.com/jmlee/curator"
Documentation = "https://github.com/jmlee/curator/blob/main/DESIGN.md"
Issues = "https://github.com/jmlee/curator/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
curator = ["py.typed", "storage/schema_v1.sql"]

# === Tooling configuration ===

[tool.pytest.ini_options]
testpaths = ["tests"]
# asyncio_mode is omitted until pytest-asyncio is installed and we have
# async tests (Phase Beta). It currently emits a PytestConfigWarning.
# Default invocation skips:
#   * tests/perf (Phase Beta v0.15 benchmarks; opt-in via ``-m slow``)
#   * any test marked ``@pytest.mark.slow``.
addopts = "-ra --strict-markers --ignore=tests/perf -m 'not slow'"
markers = [
    "slow: marks tests as slow (excluded from default run; perf benchmarks)",
    "integration: marks tests as integration tests",
    "property: marks tests as property-based",
]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line length handled by formatter
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]  # allow assert in tests

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
files = ["src"]
